requirements.txt
google-genai>=0.8.0
dnspython>=2.4.0
requests>=2.31.0
colorama>=0.4.6
python-dotenv>=1.0.0
python-whois>=0.9.4
tldextract>=5.1.1
beautifulsoup4>=4.12.0
lxml>=5.1.0
